Renamed the patrons and events connector to make it clearer that it's separate from the shared component.
